refactor adjacent items function
